add more prebuilt bindings

+++ couchbase-sys/build.rs
@@ -5,7 +5,6 @@ extern crate pkg_config;
 extern crate cmake;
 
 use std::env;
-use std::path::Path;
 use std::fs;
 
 #[cfg(feature = "build-lcb")]
@@ -32,7 +31,7 @@ fn generate_binding(bindgen_path: &str, out_dir: &str) {
         .generate_comments(false)
         .generate()
         .unwrap()
-        .write_to_file(Path::new(out_dir).join("bindings.rs"));
+        .write_to_file(std::path::Path::new(out_dir).join("bindings.rs"));
 }
 
 #[cfg(not(feature = "generate-binding"))]
@@ -41,21 +40,24 @@ fn generate_binding(_bindgen_path: &str, _out_dir: &str) {
 }
 
 fn main() {
-    let lcb_dir = "libcouchbase-2.7.4";
+    let mut version = String::from("2.7.4");
+    let lcb_dir = format!("libcouchbase-{}", version);
     let out_dir = env::var("OUT_DIR").unwrap();
 
     let mut bindgen_path = None;
     let pkg_success = if cfg!(feature = "build-lcb") {
         false
     } else {
-        let result = pkg_config::Config::new().atleast_version("2.7.4").probe("libcouchbase");
+        let result = pkg_config::Config::new().atleast_version(&version).probe("libcouchbase");
         if result.is_ok() {
-            bindgen_path = Some(String::from(result.unwrap()
+            bindgen_path = Some(String::from(result.as_ref()
+                .unwrap()
                 .include_paths
                 .get(0)
                 .expect("Could not find include path from pkg config")
                 .to_str()
                 .unwrap()));
+            version = result.unwrap().version;
             true
         } else {
             false
@@ -64,24 +66,28 @@ fn main() {
 
     if !pkg_success {
         if cfg!(feature = "build-lcb") {
-            bindgen_path = Some(build_lcb(lcb_dir, &out_dir))
+            bindgen_path = Some(build_lcb(&lcb_dir, &out_dir))
         } else {
-            panic!("Need to build libcouchbase, but the 'build-lcb' feature is not enabled!");
+            panic!("Need to build libcouchbase (none found), but the 'build-lcb' feature is not \
+                    enabled!");
         }
     }
 
-    // Step 2: From the headers, generate the rust binding via bindgen
+    // Step 2: From the headers, generate the rust binding via bindgen or load the pre-gen one
     if cfg!(feature = "generate-binding") {
         match bindgen_path {
             Some(bp) => generate_binding(&bp, &out_dir),
             None => panic!("Instructed to generate binding, but no path for headers found."),
         }
     } else {
-        let src_path = format!("{}/src/bindings-{}.rs", env!("CARGO_MANIFEST_DIR"), "2.7.4");
-        println!("{}", src_path);
+        let src_path = format!("{}/src/bindings-{}.rs",
+                               env!("CARGO_MANIFEST_DIR"),
+                               &version);
+        if !std::path::Path::new(&src_path).exists() {
+            panic!("No binding found for libcouchbase version {} in the Rust SDK source",
+                   version);
+        }
         let dst_path = format!("{}/bindings.rs", out_dir);
         fs::copy(&src_path, &dst_path).unwrap();
-        // copy the pre-generated binding in place if it can be found
-        // if it can't be found print error to say please run generate-binding feature
     }
 }
